% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datalight.R
\name{as_datalight}
\alias{as_datalight}
\title{Convert a Datafaro's datalight list object to a datalight object}
\usage{
as_datalight(data)
}
\arguments{
\item{data}{A Datafaro's datalight list object}
}
\value{
a datalight object
}
\description{
Convert a Datafaro's datalight list object to a datalight object
}
\examples{
\dontrun{
  .res <- list(
      data = mtcars,
       dictionary = list(),
       metadata = list(next_update = Sys.Date() + 30),
       cached = Sys.Date(),
       slug = "mtcars"
     )
  mi_tibble <- as_datalight(.res)
  print(mi_tibble)
}
}
