% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{auth}
\alias{auth}
\title{Authenticate and Retrieve Access Token}
\usage{
auth()
}
\value{
Does not explicitly return a value, but stores the obtained access token for future use.
}
\description{
This function prompts the user for their username and password to authenticate against an external service.
If authentication is successful, it generates and stores an access token for future requests.
}
\examples{
\dontrun{
  auth()  # Follow the prompts in the console to authenticate
}
}
