% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Banco-Central-Republica-Dominicana.R
\name{balance_isi_activos_me}
\alias{balance_isi_activos_me}
\title{Balance sectorial de las OSD: Instrumentos y sectores institucionales (Activos en ME)}
\usage{
balance_isi_activos_me(indicador = NULL, metadata = FALSE)
}
\arguments{
\item{indicador}{Vea \code{\link{downloader}}}

\item{metadata}{indica si se retornan los datos o la metadata del indicador}
}
\value{
\link{data.frame}: los datos del indicador en forma tabular
}
\description{
\lifecycle{experimental}
}
\examples{
\dontrun{
balance_isi_activos_me()
}
}
