% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Banco Central Republica Dominicana.R
\name{balance_osd_resumen}
\alias{balance_osd_resumen}
\title{Balance sectorial de las OSD: Resumen por instrumentos}
\usage{
balance_osd_resumen(indicador = NULL, metadata = FALSE)
}
\arguments{
\item{indicador}{Vea \code{\link{downloader}}}

\item{metadata}{indica si se retornan los datos o la metadata del indicador}
}
\value{
\link{data.frame}: los datos del indicador en forma tabular
}
\description{
\lifecycle{experimental}
}
\examples{
\dontrun{
  balance_osd_resumen()
}
}
