% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Banco-Central-Republica-Dominicana.R
\name{balanza_pagos_anual}
\alias{balanza_pagos_anual}
\title{Balanza de pagos anual}
\usage{
balanza_pagos_anual(indicador = NULL, metadata = FALSE)
}
\arguments{
\item{indicador}{Vea \code{\link{downloader}}}

\item{metadata}{indica si se retornan los datos o la metadata del indicador}
}
\value{
\link{data.frame}: los datos del indicador en forma tabular
}
\description{
\lifecycle{experimental}
}
\examples{
\dontrun{
balanza_pagos_anual()
}
}
