% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Banco Central Republica Dominicana.R
\name{balanza_pagos_trim}
\alias{balanza_pagos_trim}
\title{Balanza de pagos Trimestral}
\usage{
balanza_pagos_trim(indicador = NULL)
}
\arguments{
\item{indicador}{Vea \code{\link{downloader}}}
}
\value{
\link{data.frame}: los datos del indicador en forma tabular
}
\description{
\lifecycle{experimental}
}
\examples{
\dontrun{
  bpt <- balanza_pagos_trim()
}
}
