% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{db_storage}
\alias{db_storage}
\title{Almacena los datos en una base de datos}
\usage{
db_storage(db_conn, id)
}
\arguments{
\item{db_conn}{una conexión a la base de datos donde se almacenarán los datos
y donde está disponible la lista de indicadores con toda la información necesaria}

\item{id}{el ID del indicador, que debe coincidir con:
\itemize{
\item{el ID asignado en el listado de indicadores}
\item{el nombre de la función que genera los datos}
}}
}
\value{
un mensaje que indica el resultado de la operación
}
\description{
Almacena los datos en una base de datos
}
\examples{
\dontrun{
db_storage(db_conn, 'imae_mensual')
}
}
