% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Banco Central Republica Dominicana.R
\name{deuda_publica_2020}
\alias{deuda_publica_2020}
\title{Deuda pública 2020}
\usage{
deuda_publica_2020(indicador)
}
\arguments{
\item{indicador}{Vea \code{\link{downloader}}}
}
\value{
\link{data.frame}: los datos del indicador en forma tabular
}
\description{
\lifecycle{experimental}
}
\examples{
\dontrun{
  dp20 <- deuda_publica_2020()
}
}
