% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domar_list.R
\name{domar_list}
\alias{domar_list}
\title{DOMAR indicators list}
\usage{
domar_list()
}
\value{
data.frame corresponding the list of variables in domar
}
\description{
DOMAR indicators list
}
\examples{
\dontrun{
dmr_list <- domar_list()
}
}
