% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{downloader}
\alias{downloader}
\title{Descarga de archivos}
\usage{
downloader(indicador)
}
\arguments{
\item{indicador}{el vector de datos correspondiente al indicador en cuestion}
}
\value{
la ruta de un archivo temporal que contiene el archivo descargado
}
\description{
Descarga de archivos
}
\examples{
\dontrun{
downloader(pib_gasto)
}
}
