% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Banco Central Republica Dominicana.R
\name{estado_operaciones_spnf}
\alias{estado_operaciones_spnf}
\title{Estado de Operaciones del sector público no financiero (en millones de RD$)}
\usage{
estado_operaciones_spnf(indicador = NULL, metadata = FALSE)
}
\arguments{
\item{indicador}{Vea \code{\link{downloader}}}

\item{metadata}{indica si se retornan los datos o la metadata del indicador}
}
\value{
\link{data.frame}: los datos del indicador en forma tabular
}
\description{
\lifecycle{experimental}
}
\examples{
\dontrun{
estado_operaciones_spnf()
}
}
