% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{from_python}
\alias{from_python}
\title{Ejecutar funciones desde Python}
\usage{
from_python(keys, values, data = NULL, metadata = FALSE)
}
\arguments{
\item{keys}{\link{character}: keys del diccionario python que se utilizan como nombres del vector.}

\item{values}{values del diccionario python que se utilizan como el contenido del vector.}
}
\value{
Un vector con nombres que se pasa a la función del indicador.
}
\description{
Recibe los keys y valores de un diccionario python y lo convierte en un
vector con nombres en R, que luego pasa a la función del indicador.
}
