% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info.R
\name{get_metadata}
\alias{get_metadata}
\title{Obtener los metadatos de un conjunto de datos}
\usage{
get_metadata(id, .token = Sys.getenv("DATAFARO_TOKEN"))
}
\arguments{
\item{id}{\link{character} Código del conjunto de datos}

\item{.token}{\link{character} Token de acceso a la API. Por defecto se utiliza el
token guardado en la variable de entorno \code{DATAFARO_TOKEN}. Si no se ha
guardado un token, se debe especificar el token manualmente.}
}
\value{
Una lista con los metadatos del conjunto de datos. La lista tiene los siguientes
elementos:
\itemize{
\item \code{metadata} Una sublista con los metadatos de cada columna. El nombre
del elemento es el nombre de la columna. La sublista tiene los siguientes elementos:
\itemize{
\item \code{name} Nombre de la columna bien escrito
\item \code{dtype} Tipo de dato. Esto también se utiliza para el formato de datos:
\itemize{
\item \code{f2} Número flotante con 2 decimales
}
}
\item \code{notes} Un vector de caracteres con las notas del conjunto de datos.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
Esta función se utiliza junto con \code{\link{get_data}} para obtener
los metadatos como un atributo del marco de datos. Por lo tanto, raramente necesitarás
usar esta función directamente.
}
\examples{
\dontrun{
get_metadata("imae")
}
}
