% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsheet_storage.R
\name{gsheet_storage}
\alias{gsheet_storage}
\title{Storage domar data in Google Sheets}
\usage{
gsheet_storage(id, kvars, bound)
}
\arguments{
\item{id}{character ID of variable in the domar data list}

\item{kvars}{key variables for data comparison. See \code{Details}}

\item{bound}{numeric a number between 0 and 1. Default to 0.95.
Indicate the min fraction of rows where the key variables need to match.}
}
\description{
Storage domar data in Google Sheets
}
\details{
\code{dmr_list} list of variables available in domar.

Through \code{id} is selected the current variable key info.

If \code{\link{Sys.time}} greather than \code{Próxima actualización}
 See \url{https://docs.google.com/spreadsheets/d/1svANACeWShYC--wm7wwSKaUn_bF_xXuF81SnV8yMeb8/edit#gid=0}
through the \code{id} the corresponding function is executed, returning the
\code{newData}, and the \code{oldData} is retrieve from Google Sheets.

If \code{\link{nrow}} of \code{oldData} is 0. Write \code{newData} to
Google Sheet. The needed arguments are presents in indicator information.

Else both \code{newData} and \code{oldData} are subsetted by all \code{kvars}.
Both dataset are merge by the first \code{n-1} \code{kvars}. The last
variable in \code{kvars} is compared in both datasets.

\code{matchRows} The fraction of rows where the above comparison is \code{TRUE}.
\code{matchNames} The fraction of \code{\link{names}} of \code{oldData}
present in \code{\link{names}} of \code{newData}

If \code{matchNames} not iqual to 1 return an error. See \code{\link{stop}}.

Else if \code{matchRows} iqual to 1, that means both datasets are iquals,
the data is not yet updated. Then \code{dmr_list} is updated.
And next try is set to:
\itemize{
\item \code{\link{Sys.time}} plus 0.5 hours for daily frecuency.
\item \code{\link{Sys.time}} plus 4 hours for monthly frecuency.
\item \code{\link{Sys.time}} plus 20 hours for others frecuencies.
}

Else if \code{matchRows} less than \code{bound} return an error.
See \code{\link{stop}}.

Else \code{newData} is writed to Google Sheet. . The needed arguments are
presents in indicator information. Then \code{dmr_list} is updated.
And next try is set to:
\itemize{
\item \code{\link{Sys.time}} plus 23 hours for daily frecuency.
\item \code{\link{Sys.time}} plus 30 days for monthly frecuency.
\item \code{\link{Sys.time}} plus 92 days for quarter frecuency.
\item \code{\link{Sys.time}} plus 365 days for annual frecuency.
}
}
\examples{
\dontrun{
gsheet_storage(id = 'imae_month', kvars = c('fecha', 'serie', 'Indice'))
}
}
