% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imae_mensual.R
\name{imae_mensual}
\alias{imae_mensual}
\title{IMAE Mensual}
\usage{
imae_mensual(indicador)
}
\arguments{
\item{indicador}{Vea \code{\link{downloader}}}
}
\value{
\link{data.frame}: los datos del IMAE en forma tabular
}
\description{
IMAE Mensual
}
\examples{
\dontrun{
imae <- imae_mensual(indicador)
}
}
