% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imae_month.R
\name{imae_month}
\alias{imae_month}
\title{Monthly IMAE}
\usage{
imae_month(url)
}
\arguments{
\item{url}{the url of the original file in \code{dmr_list}. See \code{\link{domar_list}}}
}
\value{
data.frame data of monthly IMAE
}
\description{
Monthly IMAE
}
\details{
\code{kvars}
\itemize{
\item \code{fecha}
\item \code{serie}
\item \code{Indice}
}
}
\examples{
\dontrun{
imae <- imae_month()
}
}
