% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Banco Central Republica Dominicana.R
\name{indicadores_monetarios_bcrd}
\alias{indicadores_monetarios_bcrd}
\title{Indicadores Monetarios BCRD}
\usage{
indicadores_monetarios_bcrd(indicador)
}
\arguments{
\item{indicador}{Vea \code{\link{downloader}}}
}
\value{
\link{data.frame}: los datos del indicador en forma tabular
}
\description{
\lifecycle{experimental}
}
\examples{
\dontrun{
  imbcrd <- indicadores_monetarios_bcrd()
}
}
