% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Banco Central Republica Dominicana.R
\name{indicadores_monetarios_osd}
\alias{indicadores_monetarios_osd}
\title{Indicadores Monetarios OSD}
\usage{
indicadores_monetarios_osd(indicador)
}
\arguments{
\item{indicador}{Vea \code{\link{downloader}}}
}
\value{
\link{data.frame}: los datos del indicador en forma tabular
}
\description{
\lifecycle{experimental}
}
\examples{
\dontrun{
  imosd <- indicadores_monetarios_osd()
}
}
