% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Banco-Central-Republica-Dominicana.R
\name{indicadores_osd}
\alias{indicadores_osd}
\title{Indicadores Armonizados OSD (Activos y Pasivos Externos, Pr\\u00E9stamos y Dep\\u00F3sitos)}
\usage{
indicadores_osd(indicador = NULL, metadata = FALSE)
}
\arguments{
\item{indicador}{Vea \code{\link{downloader}}}

\item{metadata}{indica si se retornan los datos o la metadata del indicador}
}
\value{
\link{data.frame}: los datos del indicador en forma tabular
}
\description{
\lifecycle{experimental}
}
\examples{
\dontrun{
indicadores_osd()
}
}
