% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Banco Central Republica Dominicana.R
\name{inflacion_promedio_12_meses}
\alias{inflacion_promedio_12_meses}
\title{Inflación promedio 12 meses}
\usage{
inflacion_promedio_12_meses(indicador = NULL, metadata = FALSE)
}
\arguments{
\item{indicador}{Vea \code{\link{downloader}}}

\item{metadata}{indica si se retornan los datos o la metadata del indicador}
}
\value{
\link{data.frame}: los datos del indicador en forma tabular
}
\description{
\lifecycle{experimental}
}
\examples{
\dontrun{
inflacion_promedio_12_meses()
}
}
