% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDEEE.R
\name{informe_desempeno_sector_energetico}
\alias{informe_desempeno_sector_energetico}
\title{Informe de desempe\\u00F1o del sector energ\\u00E9tico}
\usage{
informe_desempeno_sector_energetico(indicador = NULL)
}
\arguments{
\item{indicador}{Vea \code{\link{downloader}}}
}
\value{
\link{data.frame}: los datos del indicador en forma tabular
}
\description{
\lifecycle{experimental}
}
\examples{
\dontrun{
  informe_desempeno_sector_energetico()
}
}
