% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDEEE.R
\name{informe_desempeno_sector_energetico_edes}
\alias{informe_desempeno_sector_energetico_edes}
\title{Informe de desempeño del sector energético: EDE's}
\usage{
informe_desempeno_sector_energetico_edes(indicador = NULL, metadata = FALSE)
}
\arguments{
\item{indicador}{Vea \code{\link{downloader}}}

\item{metadata}{indica si se retornan los datos o la metadata del indicador}
}
\value{
\link{data.frame}: los datos del indicador en forma tabular
}
\description{
\lifecycle{experimental}
}
\examples{
\dontrun{
  informe_desempeno_sector_energetico_edes()
}
}
