% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info.R
\name{list_datasets}
\alias{list_datasets}
\title{Obtener una lista de los conjuntos de datos disponibles de DataFaro}
\usage{
list_datasets(
  dataset = NULL,
  indicator = NULL,
  area = NULL,
  institution = NULL,
  freq = NULL,
  .token = Sys.getenv("DATAFARO_TOKEN")
)
}
\arguments{
\item{dataset}{\link{character} Código del conjunto de datos (opcional)}

\item{indicator}{\link{character} Código del indicador (opcional)}

\item{area}{\link{character} Código del área (opcional)}

\item{institution}{\link{character} Código de la institución (opcional)}

\item{freq}{\link{character} Código de la frecuencia
\itemize{
\item \code{H} Horaria
\item \code{D} Diaria
\item \code{W} Semanal
\item \code{M} Mensual
\item \code{Y} Anual
}}

\item{.token}{\link{character} Token de acceso a la API. Por defecto se utiliza el
token guardado en la variable de entorno \code{DATAFARO_TOKEN}. Si no se ha
guardado un token, se debe especificar el token manualmente.}
}
\value{
Un data.frame con la lista de conjuntos de datos. El data.frame tiene
las siguientes columnas:
\itemize{
\item \code{id} Código del conjunto de datos
\item \code{indicator} Código del indicador
\item \code{name} Nombre del conjunto de datos
\item \code{institution} Código de la institución
\item \code{description} Descripción del conjunto de datos
\item \code{status} Estado del conjunto de datos
\itemize{
\item \code{A} Activo
\item \code{I} Inactivo
\item \code{D} Descontinuado
\item \code{R} Reemplazado
}
\item \code{freq} Código de la frecuencia
\itemize{
\item \code{H} Horaria
\item \code{D} Diaria
\item \code{W} Semanal
\item \code{M} Mensual
\item \code{Y} Anual
}
\item \code{interval} El número de veces por frecuencia que el conjunto de
datos es actualizado
\item \code{last_update} Fecha de la última actualización
\item \code{next_update} Fecha de la próxima actualización
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
\dontrun{
list_datasets()

list_datasets(dataset = "imae")

list_datasets(area = "ECONREAL")
}
}
