% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info.R
\name{list_indicators}
\alias{list_indicators}
\title{Obtener una lista de los indicadores disponibles de DataFaro}
\usage{
list_indicators(
  indicator = NULL,
  area = NULL,
  institution = NULL,
  freq = NULL,
  .token = Sys.getenv("DATAFARO_TOKEN")
)
}
\arguments{
\item{indicator}{\link{character} Código del indicador (opcional)}

\item{area}{\link{character} Código del área (opcional)}

\item{institution}{\link{character} Código de la institución (opcional)}

\item{freq}{\link{character} Código de la frecuencia
\itemize{
\item \code{H} Horaria
\item \code{D} Diaria
\item \code{W} Semanal
\item \code{M} Mensual
\item \code{Y} Anual
}}

\item{.token}{\link{character} Token de acceso a la API. Por defecto se utiliza el
token guardado en la variable de entorno \code{DATAFARO_TOKEN}. Si no se ha
guardado un token, se debe especificar el token manualmente.}
}
\value{
Un data.frame con la lista de indicadores. El data.frame tiene
las siguientes columnas:
\itemize{
\item \code{id} Código del indicador
\item \code{name} Nombre del indicador
\item \code{description} Descripción del indicador
\item \code{area} Código del área
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
\dontrun{
list_indicators()

list_indicators(indicator = "imae")

list_indicators(area = "ECONREAL")
}
}
