% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info.R
\name{list_institutions}
\alias{list_institutions}
\title{Obtener una lista de las instituciones disponibles de DataFaro}
\usage{
list_institutions(.token = Sys.getenv("DATAFARO_TOKEN"))
}
\arguments{
\item{.token}{\link{character} Token de acceso a la API. Por defecto se utiliza el
token guardado en la variable de entorno \code{DATAFARO_TOKEN}. Si no se ha
guardado un token, se debe especificar el token manualmente.}
}
\value{
Un data.frame con la lista de instituciones. El data.frame tiene
las siguientes columnas:
\itemize{
\item \code{code} Código de la institución
\item \code{name} Nombre de la institución
\item \code{description} Descripción de la institución
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
\dontrun{
list_institutions()
}
}
