% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{log_out}
\alias{log_out}
\title{Cerrar sesión y eliminar el token de acceso}
\usage{
log_out(scope = "user")
}
\arguments{
\item{scope}{\link{character} Alcance de la autenticación para la que se desea cerrar sesión.
Por defecto es \code{"user"}. Las opciones válidas son \code{"user"}
y \code{"project"}. \code{"user"} afecta todas las sesiones de R,
mientras que \code{"project"} solo afecta a la sesión de R en el
proyecto actual.}
}
\value{
No retorna valores. Invoca un mensaje que informa al usuario que la sesión
ha sido cerrada correctamente.
}
\description{
Esta función elimina la información de autenticación almacenada para el alcance especificado,
efectivamente cerrando la sesión del usuario. Esto es útil para garantizar que la información
sensible no permanezca accesible en sesiones no autenticadas posteriormente.
}
\examples{
\dontrun{
log_out() # Ejecuta esta funci\u00f3n para cerrar la sesión y eliminar el token de acceso.
}
}
