% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{monedas_3d_codes}
\alias{monedas_3d_codes}
\title{Tipo de cambio monedas convertibles}
\format{
\link{data.frame}
}
\usage{
monedas_3d_codes
}
\description{
Tipo de cambio monedas convertibles
}
\keyword{datasets}
