% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nextTry}
\alias{nextTry}
\title{Actualiza la fecha de la próxima actualización}
\usage{
nextTry(db_conn, indicator, result)
}
\arguments{
\item{db_conn}{conexión a la base de datos donde se almacena el listado de
indicadores}

\item{indicator}{el registro de la tabla de indicadores correspondiente,
al ID del indicador en cuestión}

\item{result}{un valor 0 o 1 que indica si la actualización fué correcta o no}
}
\description{
Basado en el resultado de la actualización y la periodicidad del indicador,
actualiza la fecha de la próxima actualización.
Esta función se utiliza dentro de \code{\link{db_storage}}
}
\examples{
\dontrun{
nextTry(db_conn, indicator, 1)
}
}
