% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niveles.R
\docType{data}
\name{nvl_balanza_pagos_anual}
\alias{nvl_balanza_pagos_anual}
\title{Ordenes y niveles de los indicadores}
\format{
\link{data.frame}
}
\usage{
nvl_balanza_pagos_anual
}
\description{
Ordenes y niveles de los indicadores
}
\keyword{datasets}
