% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{nvl_indicadores_osd}
\alias{nvl_indicadores_osd}
\title{Ordenes y niveles de los indicadores}
\format{
\link{data.frame}
}
\usage{
nvl_indicadores_osd
}
\description{
Ordenes y niveles de los indicadores
}
\keyword{datasets}
