% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{nvl_pib_gasto}
\alias{nvl_pib_gasto}
\title{Ordenes y niveles de los indicadores}
\format{
\link{data.frame}

\link{data.frame}
}
\usage{
nvl_pib_gasto

nvl_pib_gasto
}
\description{
Ordenes y niveles de los indicadores

Ordenes y niveles de los indicadores
}
\keyword{datasets}
