% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{nvl_saldo_deuda_spnf_xacre}
\alias{nvl_saldo_deuda_spnf_xacre}
\title{Ordenes y niveles de los indicadores}
\format{
\link{data.frame}
}
\usage{
nvl_saldo_deuda_spnf_xacre
}
\description{
Ordenes y niveles de los indicadores
}
\keyword{datasets}
