% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Banco-Central-Republica-Dominicana.R
\name{panorama_psd}
\alias{panorama_psd}
\title{Panorama Sociedades de Dep\\u00F3sitos}
\usage{
panorama_psd(indicador = NULL, metadata = FALSE)
}
\arguments{
\item{indicador}{Vea \code{\link{downloader}}}

\item{metadata}{indica si se retornan los datos o la metadata del indicador}
}
\value{
\link{data.frame}: los datos del indicador en forma tabular
}
\description{
\lifecycle{experimental}
}
\examples{
\dontrun{
panorama_psd()
}
}
