% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Banco Central Republica Dominicana.R
\name{pib_deflactor}
\alias{pib_deflactor}
\title{Deflactor del PIB}
\usage{
pib_deflactor(indicador)
}
\arguments{
\item{indicador}{Vea \code{\link{downloader}}}
}
\value{
\link{data.frame}: los datos como se descargan del banco con muy pocas
modificaciones. Luego esos datos son tomados por pib_deflactor_trimestral y
pib_deflactor_anual
}
\description{
\lifecycle{experimental}
}
\examples{
\dontrun{
  def <- pib_deflactor()
}
}
