% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Banco Central Republica Dominicana.R
\name{pib_deflactor_anual}
\alias{pib_deflactor_anual}
\title{Deflactor del PIB Anual}
\usage{
pib_deflactor_anual(indicador = NULL)
}
\arguments{
\item{indicador}{Vea \code{\link{downloader}}}
}
\value{
\link{data.frame}: los datos del indicador en forma tabular
}
\description{
\lifecycle{experimental}
}
\examples{
\dontrun{
  dpa <- pib_deflactor_anual()
}
}
