% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Banco Central Republica Dominicana.R
\name{pib_gasto_anual}
\alias{pib_gasto_anual}
\title{PIB por enfoque del gasto anual}
\usage{
pib_gasto_anual(data = NULL, metadata = FALSE)
}
\arguments{
\item{data}{datos del indicador principal}

\item{metadata}{indica si se retornan los datos o la metadata del indicador}
}
\value{
data.frame con los datos del indicador en forma tabular
}
\description{
\lifecycle{experimental}
}
\examples{
\dontrun{
pib_gasto_anual()
}
}
