% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pib_gasto_t.R
\name{pib_gasto_t}
\alias{pib_gasto_t}
\title{PIB por enfoque del gasto trimestral}
\usage{
pib_gasto_t(indicador)
}
\arguments{
\item{indicador}{vector de datos para este indicador en la lista de domar}
}
\value{
data.frame los datos del PIB por el enfoque del gasto trimestral
}
\description{
PIB por enfoque del gasto trimestral
}
\examples{
\dontrun{
pib_gasto_t(indicador)
}
}
