% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Banco Central Republica Dominicana.R
\name{pib_gasto_trim}
\alias{pib_gasto_trim}
\title{PIB por enfoque del gasto trimestral}
\usage{
pib_gasto_trim(indicador = c(original_url = NULL, file_ext = "xls"))
}
\arguments{
\item{indicador}{vector de datos para este indicador en la lista de domar}
}
\value{
data.frame los datos del PIB por el enfoque del gasto trimestral
}
\description{
\lifecycle{experimental}
}
\examples{
\dontrun{
pib_gasto_t(indicador)
}
}
