% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Banco Central Republica Dominicana.R
\name{pib_ive_trimestral}
\alias{pib_ive_trimestral}
\title{Índices de Valores Encadenados del PIB Trimestral}
\usage{
pib_ive_trimestral(indicador = NULL)
}
\arguments{
\item{indicador}{Vea \code{\link{downloader}}}
}
\value{
\link{data.frame}: los datos del indicador en forma tabular
}
\description{
\lifecycle{experimental}
}
\examples{
\dontrun{
  ivet <- pib_ive_trimestral()
}
}
