% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Banco Central Republica Dominicana.R
\name{pib_origen_anual}
\alias{pib_origen_anual}
\title{PIB por enfoque de origen anual}
\usage{
pib_origen_anual(data = NULL, metadata = FALSE)
}
\arguments{
\item{metadata}{indica si se retornan los datos o la metadata del indicador}

\item{indicador}{Vea \code{\link{downloader}}}
}
\value{
\link{data.frame}: los datos del indicador en forma tabular
}
\description{
\lifecycle{experimental}
}
\examples{
\dontrun{
pib_origen_anual()
}
}
