% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pib_origen_trim.R
\name{pib_origen_trim}
\alias{pib_origen_trim}
\title{PIB por el enfoque de origen}
\usage{
pib_origen_trim(indicador = c(original_url = NULL, file_ext = "xlsx"))
}
\arguments{
\item{indicador}{vector con los datos del indicador}
}
\value{
los datos en formato tabular
}
\description{
PIB por el enfoque de origen
}
