% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Banco Central Republica Dominicana.R
\name{poblacion_ocupada_ingresos_nivel_educativo}
\alias{poblacion_ocupada_ingresos_nivel_educativo}
\title{Población ocupada perceptora de ingresos según nivel educativo}
\usage{
poblacion_ocupada_ingresos_nivel_educativo(indicador = NULL, metadata = FALSE)
}
\arguments{
\item{indicador}{Vea \code{\link{downloader}}}

\item{metadata}{indica si se retornan los datos o la metadata del indicador}
}
\value{
\link{data.frame}: los datos del indicador en forma tabular
}
\description{
\lifecycle{experimental}
}
\examples{
\dontrun{
poblacion_ocupada_ingresos_nivel_educativo()
}
}
