% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Banco Central Europeo.R
\name{pound_sterling_mensual_ecb}
\alias{pound_sterling_mensual_ecb}
\title{Tipo de cambio libra esterlina}
\usage{
pound_sterling_mensual_ecb(indicador = NULL)
}
\arguments{
\item{indicador}{Vea \code{\link{downloader}}}
}
\value{
\link{data.frame}: los datos del indicador en forma tabular
}
\description{
\lifecycle{experimental}
}
\examples{
\dontrun{
  pound <- pound_sterling_mensual_ecb()
}
}
