% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datalight.R
\name{print.datalight}
\alias{print.datalight}
\title{Print a datalight object}
\usage{
\method{print}{datalight}(x, ...)
}
\arguments{
\item{x}{A Datafaro's datalight json object}

\item{...}{Other arguments passed on to methods}
}
\value{
a print of the object. Invisible x
}
\description{
Print a datalight object
}
\examples{
\dontrun{
  .res <- jsonlite::toJSON(list(data = mtcars, dictionary = list(), metadata = list(next_update = Sys.Date() + 30), cached = Sys.Date(), slug = "mtcars"))
  mi_tibble <- as_datalight(.res)
  print(mi_tibble)
}
}
