% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Banco Central Republica Dominicana.R
\name{resumen_indicadores_mercado_laboral}
\alias{resumen_indicadores_mercado_laboral}
\title{Resumen principales indicadores del Mercado Laboral}
\usage{
resumen_indicadores_mercado_laboral(indicador = NULL, metadata = FALSE)
}
\arguments{
\item{indicador}{Vea \code{\link{downloader}}}

\item{metadata}{indica si se retornan los datos o la metadata del indicador}
}
\value{
\link{data.frame}: los datos del indicador en forma tabular
}
\description{
\lifecycle{experimental}
}
\examples{
\dontrun{
  resumen_indicadores_mercado_laboral()
}
}
