% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Credito Publico.R
\name{saldo_deuda_spnf_xacre}
\alias{saldo_deuda_spnf_xacre}
\title{Deuda del Sector Público No Financiero por Acreedor}
\usage{
saldo_deuda_spnf_xacre(indicador = NULL, metadata = FALSE)
}
\arguments{
\item{indicador}{Vea \code{\link{downloader}}}

\item{metadata}{indica si se retornan los datos o la metadata del indicador}
}
\value{
\link{data.frame}: los datos del indicador en forma tabular
}
\description{
\lifecycle{experimental}
}
\examples{
\dontrun{
  saldo_deuda_spnf_xacre()
}
}
