% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{secure_dbs}
\alias{secure_dbs}
\title{db_storage a prueba de errores}
\usage{
secure_dbs(db_conn, id, to_email, creds_file_name)
}
\arguments{
\item{db_conn}{vea \code{\link{db_storage}}}

\item{id}{vea \code{\link{db_storage}}}

\item{to_email}{dirección a la que enviar el email si un error ocurre.}

\item{creds_file_name}{un archivo con las credenciales necesarias para enviar emails
vea vea \url{https://rstudio.github.io/blastula/reference/create_smtp_creds_file.html}}
}
\value{
un mensaje indicando el resultado de la operación
}
\description{
Esta función ejecuta \code{\link{db_storage}} en un tryCatch de forma que si
sucede algún error se envía un email a la dirección indicada
}
\examples{
\dontrun{
 secure_dbs(db_conn, 'imae_mensual', 'domar@gmail.com', 'creds_file_name')
}
}
