% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secure_gss.R
\name{secure_gss}
\alias{secure_gss}
\title{Secure Google Sheet storage version}
\usage{
secure_gss(to_email, creds_file_name, id, ...)
}
\arguments{
\item{to_email}{email for send the error}

\item{creds_file_name}{See \code{\link[blastula]{creds_file}}}

\item{id}{See \code{\link{gsheet_storage}}}

\item{...}{See \code{\link{gsheet_storage}}
#'}
}
\description{
This is a wraper function to tryCatch for \code{\link{gsheet_storage}}.
If error, then an email is send to specified email address.
}
\examples{
\dontrun{
secure_gss(email,
 'domar.email',
  'imae_month',
   kvars = c('fecha', 'serie', 'Indice'))
}
}
