% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Banco Central Republica Dominicana.R
\name{tasa_interes_nominales_activas_bm_mensual}
\alias{tasa_interes_nominales_activas_bm_mensual}
\title{Tasas de interés nominales activas bancos múltiples mensual 2017-2020}
\usage{
tasa_interes_nominales_activas_bm_mensual(indicador = NULL)
}
\arguments{
\item{indicador}{Vea \code{\link{downloader}}}
}
\value{
\link{data.frame}: los datos del indicador en forma tabular
}
\description{
\lifecycle{experimental}
}
\examples{
\dontrun{
  tiabm <- t_interes_activas_1720_bm()
}
}
