% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Banco-Central-Republica-Dominicana.R
\name{tasas_interes_activas_bm_2008}
\alias{tasas_interes_activas_bm_2008}
\title{Tasas de inter\\u00E9s activas nominales mensuales 2008-2012 - Banco M\\u00FAltiples}
\usage{
tasas_interes_activas_bm_2008(indicador = NULL, metadata = FALSE)
}
\arguments{
\item{indicador}{Vea \code{\link{downloader}}}

\item{metadata}{indica si se retornan los datos o la metadata del indicador}
}
\value{
\link{data.frame}: los datos del indicador en forma tabular
}
\description{
\lifecycle{experimental}
}
\examples{
\dontrun{
tasas_interes_activas_bm_2008()
}
}
