% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Banco Central Republica Dominicana.R
\name{tasas_interes_activas_bm_2013}
\alias{tasas_interes_activas_bm_2013}
\title{Tasas de interés activas nominales mensuales 2013-2016 - Banco Múltiples}
\usage{
tasas_interes_activas_bm_2013(indicador = NULL, metadata = FALSE)
}
\arguments{
\item{indicador}{Vea \code{\link{downloader}}}

\item{metadata}{indica si se retornan los datos o la metadata del indicador}
}
\value{
\link{data.frame}: los datos del indicador en forma tabular
}
\description{
\lifecycle{experimental}
}
\examples{
\dontrun{
  tasas_interes_activas_bm_2013()
}
}
