% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Banco Central Republica Dominicana.R
\name{tipo_cambio_dolar_anual}
\alias{tipo_cambio_dolar_anual}
\alias{tipo_cambio_usd_dop_anual}
\title{Tipo de cambio dólar anual}
\usage{
tipo_cambio_dolar_anual(indicador = NULL, metadata = FALSE)

tipo_cambio_usd_dop_anual(...)
}
\arguments{
\item{indicador}{Vea \code{\link{downloader}}}

\item{metadata}{indica si se retornan los datos o la metadata del indicador}
}
\value{
\link{data.frame}: los datos del indicador en forma tabular
}
\description{
\lifecycle{experimental}
}
\examples{
\dontrun{
tipo_cambio_dolar_anual()
}
}
