% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Banco Central Republica Dominicana.R
\name{tipo_cambio_dolar_mensual}
\alias{tipo_cambio_dolar_mensual}
\title{Tipo de cambio dólar mensual}
\usage{
tipo_cambio_dolar_mensual(indicador = NULL)
}
\arguments{
\item{indicador}{Vea \code{\link{downloader}}}
}
\value{
\link{data.frame}: los datos del indicador en forma tabular
}
\description{
\lifecycle{experimental}
}
\examples{
\dontrun{
  tcdm <- tipo_cambio_dolar_mensual()
}
}
