% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Banco-Central-Europeo.R
\name{tipo_cambio_gbp_eur_mensual}
\alias{tipo_cambio_gbp_eur_mensual}
\title{Tipo de cambio libra esterlina}
\usage{
tipo_cambio_gbp_eur_mensual(indicador = NULL, metadata = FALSE)
}
\arguments{
\item{indicador}{Vea \code{\link{downloader}}}

\item{metadata}{indica si se retornan los datos o la metadata del indicador}
}
\value{
\link{data.frame}: los datos del indicador en forma tabular
}
\description{
\lifecycle{experimental}
}
\examples{
\dontrun{
  pound <- tipo_cambio_gbp_eur_mensual()
}
}
