% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Banco Central Europeo.R
\name{tipo_cambio_jpy_eur_mensual}
\alias{tipo_cambio_jpy_eur_mensual}
\title{Tipo de cambio Yen Japonés /Euro Mensual}
\usage{
tipo_cambio_jpy_eur_mensual(indicador = NULL, metadata = FALSE)
}
\arguments{
\item{indicador}{Vea \code{\link{downloader}}}
}
\value{
\link{data.frame}: los datos del indicador en forma tabular
}
\description{
Tipo de cambio Yen Japonés /Euro Mensual
}
\examples{
\dontrun{
 yen <- tipo_cambio_jpy_eur_mensual()
}
}
