% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_date}
\alias{to_date}
\title{Multiple variables to unique date variable}
\usage{
to_date(data, ano = "ano", mes = "mes", dia = NULL)
}
\arguments{
\item{data}{data.frame the target dataframe}

\item{ano}{numeric the character name of variable corresponding to year}

\item{mes}{character the character name of variable corresponding to month}

\item{dia}{numeric the character name of variable corresponding to day}
}
\value{
data.frame a new dataframe with the compute variable
}
\description{
Multiple variables to unique date variable
}
\examples{
\dontrun{
df <- to_date(df)
}
}
