% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Banco Central Europeo.R
\name{us_dollar_mensual_ecb}
\alias{us_dollar_mensual_ecb}
\title{Tipo de cambio USD/Euro Mensual}
\usage{
us_dollar_mensual_ecb(indicador = NULL, metadata = FALSE)
}
\arguments{
\item{indicador}{Vea \code{\link{downloader}}}
}
\value{
\link{data.frame}: los datos del indicador en forma tabular
}
\description{
Tipo de cambio USD/Euro Mensual
}
\examples{
\dontrun{
 usd_eur <- us_dollar_mensual_ecb(indicador)
}
}
