% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_data}
\alias{validate_data}
\title{Valida los datos antes de escribir a la base de datos}
\usage{
validate_data(db_conn, newData, oldData, indicator)
}
\arguments{
\item{db_conn}{conexion a base de datos}

\item{newData}{la data obtenida con la función correspondiente}

\item{oldData}{la data que se encuentra actualmente en la base de datos}

\item{indicator}{current indicator}
}
\value{
número de filas distintas
}
\description{
Valida los datos antes de escribir a la base de datos
}
