% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Banco Central Europeo.R
\name{yen_mensual_ecb}
\alias{yen_mensual_ecb}
\title{Tipo de cambio Yen Japonés /Euro Mensual}
\usage{
yen_mensual_ecb(indicador = NULL)
}
\arguments{
\item{indicador}{Vea \code{\link{downloader}}}
}
\value{
\link{data.frame}: los datos del indicador en forma tabular
}
\description{
Tipo de cambio Yen Japonés /Euro Mensual
}
\examples{
\dontrun{
 yen <- yen_mensual_ecb()
}
}
