% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{auth}
\alias{auth}
\title{Autenticar y obtener token de acceso}
\usage{
auth()
}
\value{
Retorna el token de acceso si la autenticación es exitosa o el token almacenado sigue siendo válido.
En otros casos, guía al usuario para realizar una nueva autenticación.
}
\description{
Esta función verifica si existe información de autenticación previa almacenada y válida.
Si no encuentra información válida o el token ha expirado, solicita al usuario que se autentique nuevamente.
Si la autenticación es exitosa, genera y almacena un nuevo token de acceso para solicitudes futuras.
En caso de encontrar un token válido y no expirado, lo retorna para su uso inmediato.
}
\examples{
\dontrun{
auth() # Ejecuta esta funci\u00f3n para autenticarte o validar el token existente.
}
}
