% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balanza_pagos_anual.R
\name{balanza_pagos_anual}
\alias{balanza_pagos_anual}
\title{Balanza de pagos Anual}
\usage{
balanza_pagos_anual(indicador = NULL)
}
\arguments{
\item{indicador}{Vea \code{\link{downloader}}}
}
\value{
\link{data.frame}: los datos del indicador en forma tabular
}
\description{
Balanza de pagos Anual
}
\examples{
\dontrun{
  bpa <- balanza_pagos_anual()
}
}
