% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domar_list.R
\name{domar_list}
\alias{domar_list}
\title{DOMAR indicators list}
\usage{
domar_list()
}
\value{
data.frame corresponding the list of indicatos in domar list
}
\description{
DOMAR indicators list
}
\details{
Don't use if you don't have access to \code{dmr_list} file.
See \url{https://docs.google.com/spreadsheets/d/1svANACeWShYC--wm7wwSKaUn_bF_xXuF81SnV8yMeb8/edit#gid=0}
}
\examples{
\dontrun{
dmr_list <- domar_list()
}
}
