% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{downloader}
\alias{downloader}
\title{Download files interactivelly}
\usage{
downloader(url, ext)
}
\arguments{
\item{url}{url address of file}

\item{ext}{the file extention for the downloaded file.
See \code{\link{tempfile}}}
}
\value{
pointer to tempfile storage
}
\description{
Download files interactivelly
}
\examples{
\dontrun{
downloader('https://cdn.bancentral.gov.do/documents/
estadisticas/sector-real/documents/imae.xlsx', 'xlsx')
}
}
