% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{downloader}
\alias{downloader}
\title{Descarga de archivos}
\usage{
downloader(indicador)
}
\arguments{
\item{indicador}{el vector de datos correspondiente al indicador en cuestion.
Vea la secci\\u00F3n \code{details} para m\\u00E1s detalles.}
}
\value{
la ruta de un archivo temporal que contiene el archivo descargado
}
\description{
Descarga de archivos
}
\details{
El indicador pasado a la funci\\u00F3n es un vector de datos que debe
contener por lo menos:

\itemize{
\item{Extensi\\u00F3n del archivo \code{(file_ext)}: uno de 'xlsx', 'xls',
'csv', 'pdf', 'dir', 'sql'}
\item{Enlace directo a la descarga \code{(original_url)}}
}
}
\examples{
\dontrun{
   downloader(pib_gasto)
}
}
