% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Obtener la última actualización de un conjunto de datos de Datafaro}
\usage{
get_data(code, .token = .get_valid_token(), .cached = TRUE)
}
\arguments{
\item{code}{\link{character} Código del conjunto de datos}

\item{.token}{\link{character} Token de acceso a la API. Si no se ha
guardado un token, se debe especificar el token manualmente.}

\item{.cached}{\link{logical} Si se debe utilizar la caché. Por defecto es \code{TRUE}.}
}
\value{
Un datalight con la última actualización del conjunto de datos.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
\dontrun{
get_data("NCP_XDC")
}
}
