% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imae_mensual.R
\name{imae_mensual}
\alias{imae_mensual}
\title{IMAE Mensual}
\usage{
imae_mensual(url)
}
\arguments{
\item{url}{el url del archivo original}
}
\value{
data.frame los datos del IMAE en forma tabular
}
\description{
IMAE Mensual
}
\examples{
\dontrun{
imae <- imae_mensual('https://.../imae.xlsx')
}
}
