% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Banco Central Republica Dominicana.R
\name{imae_mensual}
\alias{imae_mensual}
\title{Indicador Mensual de Actividad Económica (IMAE) Mensual}
\usage{
imae_mensual(indicador)
}
\arguments{
\item{indicador}{Vea \code{\link{downloader}}}
}
\value{
\link{data.frame}: los datos del IMAE en forma tabular
}
\description{
\lifecycle{experimental}
}
\examples{
\dontrun{
imae <- imae_mensual(indicador)
}
}
