% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Banco Central Republica Dominicana.R
\name{ipc_mensual_2020}
\alias{ipc_mensual_2020}
\title{Índice de Precios al Consumidor (octubre 2019 - septiembre 2020)}
\usage{
ipc_mensual_2020(indicador = NULL, metadata = FALSE)
}
\arguments{
\item{indicador}{Vea \code{\link{downloader}}}

\item{metadata}{indica si se retornan los datos o la metadata del indicador}
}
\value{
\link{data.frame}: los datos del indicador en forma tabular
}
\description{
\lifecycle{experimental}
}
\examples{
\dontrun{
  ipc_mensual_2020()
}
}
