% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Banco Central Republica Dominicana.R
\name{pib_deflactor}
\alias{pib_deflactor}
\title{Deflactor del Producto Interno Bruto (PIB)}
\usage{
pib_deflactor(indicador = NULL)
}
\arguments{
\item{indicador}{Vea \code{\link{downloader}}}
}
\description{
\lifecycle{deprecated}
}
