% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Banco Central Republica Dominicana.R
\name{pib_gasto_trim}
\alias{pib_gasto_trim}
\title{PIB por enfoque del gasto trimestral}
\usage{
pib_gasto_trim(indicador = NULL, metadata = FALSE)
}
\arguments{
\item{indicador}{vector de datos para este indicador en la lista de domar}

\item{metadata}{indica si se retornan los datos o la metadata del indicador}
}
\value{
data.frame con los datos del indicador en forma tabular
}
\description{
\lifecycle{experimental}
}
\examples{
\dontrun{
pib_gasto_trim()
}
}
