% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Banco Central Republica Dominicana.R
\name{pib_ive}
\alias{pib_ive}
\title{Índices de Valores Encadenados (IVE)}
\usage{
pib_ive(indicador = NULL)
}
\arguments{
\item{indicador}{Vea \code{\link{downloader}}}
}
\description{
\lifecycle{experimental}
}
