% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/us_dollar_mensual_ecb.R
\name{us_dollar_mensual_ecb}
\alias{us_dollar_mensual_ecb}
\title{Tipo de cambio USD/Eur Mensual}
\usage{
us_dollar_mensual_ecb(
  indicador = c(original_url =
    "https://sdw.ecb.europa.eu/quickview.do?SERIES_KEY=120.EXR.M.USD.EUR.SP00.A",
    file_ext = "html")
)
}
\arguments{
\item{indicador}{Vea \code{\link{downloader}}}
}
\value{
\link{data.frame}: los datos del indicador en forma tabular
}
\description{
Tipo de cambio USD/Eur Mensual
}
\examples{
\dontrun{
 usd_eur <- us_dollar_mensual_ecb(indicador)
}
}
